requirements.txt
PyQt5~=5.15.7
PyYAML~=6.0
scapy~=2.5.0
requests~=2.28.1
wappalyzer~=1.0
python-Wappalyzer~=0.3.1